refactor to NetlistService
